+++ geostat.h
@@ -80,7 +80,7 @@ float haversin(float theta);
 float distance(float lat1, float lon1, float lat2, float lon2);
 int dlon2m(int lat);
 uint32_t dlon2mm(float lat);
-void calculate_grid(int lat);
+void calculate_grid(int8_t lat);
 void clear_storage(void);
 int store_lon_lat(float lon, float lat, float heading);
 void print_storage(void);
